Moved timer stuff to weeklyboss
